Add a demuxer thread wrapper
This is relatively similar to stream/cache.c: it wraps the existing,
single-threaded demuxer interface, and acts as a cache. It tries to
proxy most operations, but some operations still require synchronous
access. (Much like the stream cache wrapper.)

In some cases it's relatively hard to make sure the new interface will
behave as the old one. Also beware of "special" things like DVB, TV,
and some others, which don't use proper STREAM_CTRLs, and thus will
not behave as expected.

The way how the stream is wrapped is the most hacky part about this.
There are various things that still need access to the stream, but the
demuxer requires access to the stream as well, so it has to be wrapped
in some ways too.

Immediate things that are left to do:
- ordered chapters or similar aren't thread-wrapped
- add a switch to control queue depth or whether thread should be used
- do something about realtime streaming (typically requires you to
  read-ahead infinite packets, or you'll get packet drops
- probably more